move old widgets into legacy
